--- DesignScape/DesignScape/Model/Product.swift
@@ -1,18 +1,0 @@
-//
-//  Product.swift
-//  DesignScape
-//
-//  Created by Y Nguyen on 2/21/24.
-//
-
-import SwiftUI
-import Firebase
-import FirebaseFirestoreSwift
-
-struct Product: Identifiable, Codable, Hashable {
-    @DocumentID var id: String?
-    var name: String
-    var price: Double
-    var description: String
-    var imageURL: String
-}
